pyproject.toml
[project]
name = "twistr"
version = "0.1.0"
description = "Computational pipeline for α-helix-mediated protein binder design: PDB curation → atom14 tensor extraction → ML training-example assembly → epitope selection on the target antigen → lead-optimization structure prediction → Rosetta-based linker design."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Julian Alverio", email = "julian.a.alverio@gmail.com" },
]
license = { text = "Proprietary" }
keywords = [
    "protein design",
    "protein engineering",
    "structural biology",
    "machine learning",
    "computational biology",
    "binder design",
    "PPI",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Operating System :: POSIX",
]
dependencies = [
    "requests>=2.31",
    "gemmi>=0.6.5",
    "pandas>=2.1",
    "pyarrow>=15.0",
    "pyyaml>=6.0",
    "numpy>=1.26",
    "modal>=0.64",
    "scipy>=1.11",
    "freesasa>=2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
ml = [
    "torch>=2.2",
    # Was `lightning>=2.2` historically; the standalone `lightning` PyPI
    # package was unlisted ~2026-04 and pip can no longer find any 2.x
    # versions. `pytorch-lightning` provides the same `lightning.pytorch`
    # namespace, so the `import lightning.pytorch as pl` lines still work.
    "pytorch-lightning>=2.2",
    "wandb>=0.16",
    "rapidfuzz>=3.6",
]
agent = [
    "torch>=2.2",
    "pytorch-lightning>=2.2",
    "anthropic>=0.40",
]

[project.scripts]
twistr = "twistr.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["twistr"]
exclude = [
    "twistr/runtime",
    "twistr/dev",
]

[tool.pytest.ini_options]
markers = [
    "network: integration tests that hit the live RCSB/EBI APIs",
]
testpaths = ["twistr/dev/tests"]
norecursedirs = ["twistr/external", "external", "*.egg", ".venv*", ".cache", "twistr/runtime"]
